% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitstree.R
\name{splitstree}
\alias{splitstree}
\title{Convert an object to SplitsTree-compatible NEXUS input
and launch SplitsTree if required}
\usage{
splitstree(dist, nexus.file = NULL, plot = FALSE,
  splitstree.path = getOption("splitstree.path", NULL))
}
\arguments{
\item{dist}{A distance object (usually of class 'dist')}

\item{nexus.file}{A name of the file where the NEXUS file will be written to}

\item{plot}{Set this to 'PDF' or 'SVG' to automatically invoke SplitsTree4 and generate a graphics file}

\item{splitstree.path}{Path to the SplitsTree4 binary}
}
\value{
The name of the generated NEXUS file
}
\description{
Convert an object to SplitsTree-compatible NEXUS input
and launch SplitsTree if required
}
\examples{
library(cluster)
data(agriculture)
agriculture.dist <- daisy(agriculture)
splitstree(agriculture.dist, plot='PDF')
}

